style: refactor showlongPopup text with one component
